requirements.txt
alabaster==0.7.16
annotated-types==0.6.0
anytree==2.12.1
astroid==3.1.0
attrs==23.2.0
Authlib==1.3.0
autopage==0.5.2
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==24.3.0
blinker==1.7.0
bullet==2.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cliff==4.6.0
cmd2==2.4.3
colorama==0.4.6
configobj==5.0.8
configparser==6.0.1
cryptography==3.3.2
dill==0.3.8
distlib==0.3.8
dnspython==2.0.0
docutils==0.20.1
ecdsa==0.18.0
email_validator==2.1.1
filelock==3.13.3
Flask==3.0.0
Flask-Authlib==1.6.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gnupg==2.3.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-oauth2-tool==0.0.3
googleapis-common-protos==1.63.0
greenlet==3.0.1
gunicorn==21.2.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.6
imagesize==1.4.1
iniconfig==2.0.0
install==1.3.5
ipwhois==1.2.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pbr==6.0.0
pexpect==4.9.0
platformdirs==4.2.0
pluggy==1.4.0
prettytable==3.10.0
property-manager==3.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydig==0.4.0
pyflakes==3.2.0
Pygments==2.17.2
pylint==3.1.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyperclip==1.8.2
pyproject-api==1.6.1
pyroute2==0.7.12
pytest==7.4.3
python-jose==3.3.0
python-secrets==23.4.2
python-wireguard==0.2.2
pywgkey==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
setuptools-scm==8.0.4
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.29
stevedore==5.2.0
tomlkit==0.12.4
tox==4.14.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
verboselogs==1.7
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.1
wgconfig==1.0.1
xkcdpass==1.19.9
yamlreader==3.0.4
